Moved ID func to solver. Made more complex.
ID func now bit packs the deck to 39 chars,
rather than a full 52. This can save a substantial
amount of space for large tree games.

Tested to be about 12% less memory over 50m items,
while making the ID func run about 13% longer. As
this takes up the vast bulk of the program's memory,
but not the vast bulk of the execution time, this
seems like a fair trade off.