Added move priority.
Follows what appears to be generally accepted logic
for what constitutues good move order for Klondike.

In testing, there was a significant decrease in the
average number of positions tried to solve a winnable
game, and a significant increase of games that changed
from unknown to win (tested with max 10m positions).